requirements.txt
openai
python-dotenv
uuid